fixed quartz line drawing code for arm64 platforms

+++ simpleUI/Classes/PlotView.h
@@ -13,19 +13,19 @@
     float* data; // the data to plot
 	unsigned int length; // the length of data array
 	// y axis range for plot:
-	float minY;
-	float maxY;
-	float* lineColor; // RGB array
+	CGFloat minY;
+	CGFloat maxY;
+	CGFloat* lineColor; // RGB array
 }
 
 @property (nonatomic) float* data;
 @property (nonatomic) unsigned int length;
-@property (nonatomic) float minY;
-@property (nonatomic) float maxY;
-@property (nonatomic) float* lineColor;
+@property (nonatomic) CGFloat minY;
+@property (nonatomic) CGFloat maxY;
+@property (nonatomic) CGFloat* lineColor;
 
 // set the y axis range of the plot
--(void)setYRange_min: (float)Ymin  max:(float)Ymax;
+-(void)setYRange_min: (CGFloat)Ymin  max:(CGFloat)Ymax;
 // automatically set the range based on the values in the vector
 -(void)autoRange;
 -(id)initWith_Frame:(CGRect)frame;
